+++ packages/messaging/src/controllers/sw-controller.test.ts
@@ -14,55 +14,67 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { expect } from 'chai';
-import { stub, spy } from 'sinon';
 
-import { getFakeFirebaseDependencies } from '../testing/fakes/firebase-dependencies';
 import '../testing/setup';
-import { SwController, BgMessageHandler } from './sw-controller';
+
 import * as tokenManagementModule from '../core/token-management';
-import { Stub } from '../testing/sinon-types';
-import { Writable, ValueOf, DeepPartial } from 'ts-essentials';
-import { MessagePayload } from '../interfaces/message-payload';
-import { MessageType, InternalMessage } from '../interfaces/internal-message';
+
+import { BgMessageHandler, SwController } from './sw-controller';
 import {
-  mockServiceWorker,
-  restoreServiceWorker,
-  FakeEvent,
-  FakePushSubscription
-} from '../testing/fakes/service-worker';
-import {
-  FCM_MSG,
-  DEFAULT_VAPID_KEY,
+  CONSOLE_CAMPAIGN_ANALYTICS_ENABLED,
   CONSOLE_CAMPAIGN_ID,
   CONSOLE_CAMPAIGN_NAME,
   CONSOLE_CAMPAIGN_TIME,
-  CONSOLE_CAMPAIGN_ANALYTICS_ENABLED
+  DEFAULT_VAPID_KEY,
+  FCM_MSG
 } from '../util/constants';
-import { dbSet } from '../helpers/idb-manager';
-import { getFakeTokenDetails } from '../testing/fakes/token-details';
+import { DeepPartial, ValueOf, Writable } from 'ts-essentials';
+import {
+  FakeEvent,
+  FakePushSubscription,
+  mockServiceWorker,
+  restoreServiceWorker
+} from '../testing/fakes/service-worker';
+import {
+  MessagePayloadInternal,
+  MessageType
+} from '../interfaces/internal-message-payload';
+import { spy, stub } from 'sinon';
+
 import { FirebaseInternalDependencies } from '../interfaces/internal-dependencies';
+import { Stub } from '../testing/sinon-types';
+import { dbSet } from '../helpers/idb-manager';
+import { expect } from 'chai';
+import { getFakeFirebaseDependencies } from '../testing/fakes/firebase-dependencies';
+import { getFakeTokenDetails } from '../testing/fakes/token-details';
 
 // Add fake SW types.
 declare const self: Window & Writable<ServiceWorkerGlobalScope>;
 
-const NOTIFICATION_MESSAGE_PAYLOAD: MessagePayload = {
+// internal message payload (parsed directly from the push event) that contains and only contains
+// notification payload.
+const DISPLAY_MESSAGE: MessagePayloadInternal = {
   notification: {
-    title: 'message title',
-    body: 'message body',
-    data: {
-      key: 'value'
-    }
+    title: 'title',
+    body: 'body'
   },
   fcmOptions: {
     link: 'https://example.org'
-  }
+  },
+  from: 'from',
+  // eslint-disable-next-line camelcase
+  collapse_key: 'collapse'
 };
 
-const DATA_MESSAGE_PAYLOAD: MessagePayload = {
+// internal message payload (parsed directly from the push event) that contains and only contains
+// data payload.
+const DATA_MESSAGE: MessagePayloadInternal = {
   data: {
     key: 'value'
-  }
+  },
+  from: 'from',
+  // eslint-disable-next-line camelcase
+  collapse_key: 'collapse'
 };
 
 describe('SwController', () => {
@@ -78,10 +90,9 @@ describe('SwController', () => {
 
     stub(Notification, 'permission').value('granted');
 
-    // Instead of calling actual addEventListener, add the event to the
-    // eventListeners list.
-    // Actual event listeners can't be used as the tests are not running in a
-    // Service Worker, which means Push events do not exist.
+    // Instead of calling actual addEventListener, add the event to the eventListeners list. Actual
+    // event listeners can't be used as the tests are not running in a Service Worker, which means
+    // Push events do not exist.
     addEventListenerStub = stub(self, 'addEventListener').callsFake(
       (type, listener) => {
         eventListenerMap.set(type, listener);
@@ -204,16 +215,14 @@ describe('SwController', () => {
       await callEventListener(
         makeEvent('push', {
           data: {
-            json: () => NOTIFICATION_MESSAGE_PAYLOAD
+            json: () => DISPLAY_MESSAGE
           }
         })
       );
 
-      const expectedMessage: InternalMessage = {
-        firebaseMessaging: {
-          type: MessageType.PUSH_RECEIVED,
-          payload: NOTIFICATION_MESSAGE_PAYLOAD
-        }
+      const expectedMessage: MessagePayloadInternal = {
+        ...DISPLAY_MESSAGE,
+        messageType: MessageType.PUSH_RECEIVED
       };
       expect(postMessageSpy).to.have.been.calledOnceWith(expectedMessage);
     });
@@ -226,17 +235,16 @@ describe('SwController', () => {
       await callEventListener(
         makeEvent('push', {
           data: {
-            json: () => NOTIFICATION_MESSAGE_PAYLOAD
+            json: () => DISPLAY_MESSAGE
           }
         })
       );
 
       expect(postMessageSpy).not.to.have.been.called;
-      expect(showNotificationSpy).to.have.been.calledWith('message title', {
-        ...NOTIFICATION_MESSAGE_PAYLOAD.notification,
+      expect(showNotificationSpy).to.have.been.calledWith('title', {
+        ...DISPLAY_MESSAGE.notification,
         data: {
-          ...NOTIFICATION_MESSAGE_PAYLOAD.notification!.data,
-          [FCM_MSG]: NOTIFICATION_MESSAGE_PAYLOAD
+          [FCM_MSG]: DISPLAY_MESSAGE
         }
       });
     });
@@ -247,16 +255,16 @@ describe('SwController', () => {
       await callEventListener(
         makeEvent('push', {
           data: {
-            json: () => NOTIFICATION_MESSAGE_PAYLOAD
+            json: () => DISPLAY_MESSAGE
           }
         })
       );
 
-      expect(showNotificationSpy).to.have.been.calledWith('message title', {
-        ...NOTIFICATION_MESSAGE_PAYLOAD.notification,
+      expect(showNotificationSpy).to.have.been.calledWith('title', {
+        ...DISPLAY_MESSAGE.notification,
         data: {
-          ...NOTIFICATION_MESSAGE_PAYLOAD.notification!.data,
-          [FCM_MSG]: NOTIFICATION_MESSAGE_PAYLOAD
+          ...DISPLAY_MESSAGE.notification!.data,
+          [FCM_MSG]: DISPLAY_MESSAGE
         }
       });
     });
@@ -268,12 +276,37 @@ describe('SwController', () => {
       await callEventListener(
         makeEvent('push', {
           data: {
-            json: () => DATA_MESSAGE_PAYLOAD
+            json: () => DATA_MESSAGE
           }
         })
       );
 
       expect(bgMessageHandlerSpy).to.have.been.calledWith();
+    });
+
+    it('forwards MessagePayload with a notification payload to onBackgroundMessage', async () => {
+      const bgMessageHandlerSpy = spy();
+      const showNotificationSpy = spy(self.registration, 'showNotification');
+
+      swController.onBackgroundMessage(bgMessageHandlerSpy);
+
+      await callEventListener(
+        makeEvent('push', {
+          data: {
+            json: () => ({
+              notification: {
+                ...DISPLAY_MESSAGE
+              },
+              data: {
+                ...DATA_MESSAGE
+              }
+            })
+          }
+        })
+      );
+
+      expect(bgMessageHandlerSpy).to.have.been.called;
+      expect(showNotificationSpy).to.have.been.called;
     });
 
     it('warns if there are more action buttons than the browser limit', async () => {
@@ -291,7 +324,7 @@ describe('SwController', () => {
           data: {
             json: () => ({
               notification: {
-                ...NOTIFICATION_MESSAGE_PAYLOAD,
+                ...DISPLAY_MESSAGE,
                 actions: [
                   { action: 'like', title: 'Like' },
                   { action: 'favorite', title: 'Favorite' }
@@ -308,7 +341,7 @@ describe('SwController', () => {
     });
   });
 
-  describe('setBackgrounMessageHandler', () => {
+  describe('setBackgroundMessageHandler', () => {
     it('throws on invalid input', () => {
       expect(() =>
         swController.setBackgroundMessageHandler(
@@ -350,11 +383,11 @@ describe('SwController', () => {
 
     beforeEach(() => {
       NOTIFICATION_CLICK_PAYLOAD = {
-        notification: new Notification('message title', {
-          ...NOTIFICATION_MESSAGE_PAYLOAD.notification,
+        notification: new Notification('title', {
+          ...DISPLAY_MESSAGE.notification,
           data: {
-            ...NOTIFICATION_MESSAGE_PAYLOAD.notification!.data,
-            [FCM_MSG]: NOTIFICATION_MESSAGE_PAYLOAD
+            ...DISPLAY_MESSAGE.notification!.data,
+            [FCM_MSG]: DISPLAY_MESSAGE
           }
         })
       };
@@ -437,10 +470,8 @@ describe('SwController', () => {
       expect(openWindowSpy).not.to.have.been.called;
       expect(focusSpy).to.have.been.called;
       expect(postMessageSpy).to.have.been.calledWith({
-        firebaseMessaging: {
-          type: MessageType.NOTIFICATION_CLICKED,
-          payload: NOTIFICATION_MESSAGE_PAYLOAD
-        }
+        ...DISPLAY_MESSAGE,
+        messageType: MessageType.NOTIFICATION_CLICKED
       });
     });