+++ size-analysis-web-app/src/constants.jsx
@@ -28,6 +28,6 @@ export const ENDPOINTS = {
     generateSizeAnalysisReportGivenCustomBundle: 'generateSizeAnalysisReportGivenCustomBundle'
 }
 
-//export const API_ROOT = 'https://us-central1-web-app-size-analysis.cloudfunctions.net/';
-export const API_ROOT = 'http://localhost:5001/web-app-size-analysis/us-central1/';
+export const API_ROOT = 'https://us-central1-fir-size-analysis.cloudfunctions.net/';
+//export const API_ROOT = 'http://localhost:5001/fir-size-analysis/us-central1/';